+++ changelog.md
@@ -1,5 +1,13 @@
 # Pokéstar Changelog
 
+### 1.1.1
+
+- Increase spawned Pokemon duration 15 min -> 30 min
+- Fix some performance issues
+- Fix `/help` bug
+- Add vote, events to tutorial
+- Fix a weird voting bug
+
 TODO:
 
 - Stretch: New event